Created the garden
