+++ components.d.ts
@@ -7,18 +7,18 @@ export {}
 
 declare module 'vue' {
   export interface GlobalComponents {
-    BlogListItem: typeof import('./src/pages/blog/components/BlogListItem.vue')['default']
-    BlogListSearch: typeof import('./src/pages/blog/components/BlogListSearch.vue')['default']
-    BlogPostList: typeof import('./src/pages/blog/components/BlogPostList.vue')['default']
     CoreAppHeader: typeof import('./src/core/components/app/header/CoreAppHeader.vue')['default']
     CoreBtn: typeof import('./src/core/components/CoreBtn.vue')['default']
     CoreDivider: typeof import('./src/core/components/CoreDivider.vue')['default']
-    CoreFooter: typeof import('./src/core/components/CoreFooter.vue')['default']
-    CoreMdWrapper: typeof import('./src/core/components/CoreMdWrapper.vue')['default']
-    CoreMenu: typeof import('./src/core/components/CoreMenu.vue')['default']
-    CoreNavButton: typeof import('./src/core/components/CoreNavButton.vue')['default']
+    CoreFooter: typeof import('./src/core/components/app/CoreFooter.vue')['default']
+    CoreMdWrapper: typeof import('./src/core/components/containers/CoreMdWrapper.vue')['default']
+    CoreMenu: typeof import('./src/core/components/containers/CoreMenu.vue')['default']
+    CoreNavButton: typeof import('./src/core/components/app/header/CoreNavButton.vue')['default']
     CoreNavButtons: typeof import('./src/core/components/app/header/CoreNavButtons.vue')['default']
     CoreNavMenu: typeof import('./src/core/components/app/header/CoreNavMenu.vue')['default']
+    CorePostList: typeof import('./src/core/components/posts/CorePostList.vue')['default']
+    CorePostListItem: typeof import('./src/core/components/posts/CorePostListItem.vue')['default']
+    CoreSearchBar: typeof import('./src/core/components/inputs/CoreSearchBar.vue')['default']
     ProjectListItem: typeof import('./src/pages/projects/components/ProjectListItem.vue')['default']
     ProjectsList: typeof import('./src/pages/projects/components/ProjectsList.vue')['default']
     RouterLink: typeof import('vue-router')['RouterLink']
