+++ src/core/components/app/CoreFooter.vue
