+++ src/core/components/app/header/CoreNavButtons.vue
@@ -16,6 +16,9 @@ withDefaults(
       <core-btn to="/" title="Home" underline>Home</core-btn>
       <core-btn to="/blog" title="blog" underline>Blog</core-btn>
       <core-btn to="/projects" title="Projects" underline>Projects</core-btn>
+      <core-btn to="/digital-garden" title="Digital Garden" underline>
+        Garden
+      </core-btn>
     </nav>
   </div>
 </template>
