+++ src/core/components/containers/CoreMenu.vue
