+++ src/core/components/inputs/CoreSearchBar.vue
