+++ src/core/components/posts/CorePostList.vue
@@ -1,0 +1,16 @@
+<script lang="ts" setup>
+const props = defineProps<{
+  postsPath: string;
+}>();
+
+const { posts, search } = usePosts(props.postsPath);
+</script>
+
+<template>
+  <div>
+    <CoreSearchBar v-model="search" placeholder="Search..." />
+    <ul class="blog-list flex flex-col list-none gap-sm pl-none">
+      <CorePostListItem v-for="post in posts" :key="post.path" :post="post" />
+    </ul>
+  </div>
+</template>
