+++ src/pages/digital-garden/index.md
@@ -1,5 +1,7 @@
 ---
 title: Digital Garden - Freedom Evenden
 description: Freedom Evenden's digital garden
-subtitle:
+subtitle: This is my digital garden. Posts may be incomplete and fragmented. Have a look around! feel free to take some ideas back to your own digital garden.
 ---
+
+<CorePostList posts-path="/digital-garden/" />
