+++ src/pages/digital-garden/the-small-web.md
@@ -1,0 +1,16 @@
+---
+title: The Small Web
+description: My ramblings and thoughts about the small web.
+---
+
+## Overview
+
+Recently I've been exploring the small web. It's sort of an anti-corporate counter movement to the direction the web has gone over the last decade or two. These days there's only a handful of places most people tend to hangout on the internet and all of them are controlled by big companies. These large websites are usually piled with all kinds of trackers and ads, and tend to be slow on older/cheaper hardware. Additionally, some people feel that most of these large websites have a sanitized, corporate feel. They don't really feel like they're made by humans.
+
+Enter the small web. It's a decentralized (Not in the Crypto-bro way), loosely connected network of websites that harkens back to the earlier days of the internet. Mostly consisting of personal websites and blogs. Most small websites are pretty simple, consisting of mainly html and css. With little to no JavaScript. More extreme examples would be something like [Project Gemini](https://geminiprotocol.net/) or [Gopher](<https://en.wikipedia.org/wiki/Gopher_(protocol)>). They're alternative internet protocals to HTTP, but they are purely text focused, no styles, or JavaScript. You can even use them from a terminal.
+
+## My Thoughts
+
+So far I really like everything I've been learning about the small web. I think things like Project Gemini are super cool and I'm glad stuff like that exists. But, I like being able to utilize CSS and JavaScript. They allow us to build cool stuff that isn't possible on something like Gemini. Additionally you can actually access most Gemini based projects via the regular web so you can have the best of both worlds.
+
+The small web was one of my primary motivations for building this website. I like the philosophy behind it but I don't think that means we need to abandon the regular web all together and jump on something like Gemini. Instead we should use the tools like CSS and JavaScript to enrich our websites while focusing on usability and content first.
